+++ go.mod
@@ -16,7 +16,7 @@ require (
 	github.com/spf13/cast v1.5.1
 	github.com/zclconf/go-cty v1.13.3
 	go.mongodb.org/atlas v0.33.0
-	go.mongodb.org/atlas-sdk/v20230201005 v20230201005.0.1-0.20230825095504-8d6f690cf806
+	go.mongodb.org/atlas-sdk/v20230201006 v20230201006.0.0
 	go.mongodb.org/realm v0.1.0
 	golang.org/x/exp v0.0.0-20221208152030-732eee02a75a
 )
