+++ mongodbatlas/data_source_mongodbatlas_cloud_shared_tier_restore_jobs.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	atlasSDK "go.mongodb.org/atlas-sdk/v20230201005/admin"
+	atlasSDK "go.mongodb.org/atlas-sdk/v20230201006/admin"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
