+++ mongodbatlas/data_source_mongodbatlas_federated_database_instances.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"go.mongodb.org/atlas-sdk/v20230201005/admin"
+	"go.mongodb.org/atlas-sdk/v20230201006/admin"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
