+++ mongodbatlas/provider_test.go
@@ -1,9 +1,6 @@
 package mongodbatlas
 
 import (
-	"context"
-	"encoding/json"
-	"fmt"
 	"os"
 	"strings"
 	"testing"
@@ -11,9 +8,7 @@ import (
 	matlas "go.mongodb.org/atlas/mongodbatlas"
 
 	"github.com/go-test/deep"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 )
 
 const (
@@ -254,20 +249,4 @@ func testCheckPrivateEndpointServiceDataFederationOnlineArchiveRun(tb testing.TB
 	if os.Getenv("MONGODB_ATLAS_PRIVATE_ENDPOINT_ID") == "" {
 		tb.Skip("`MONGODB_ATLAS_PRIVATE_ENDPOINT_ID` must be set for Private Endpoint Service Data Federation and Online Archive acceptance testing")
 	}
-}
-
-var _ plancheck.PlanCheck = debugPlan{}
-
-type debugPlan struct{}
-
-func (e debugPlan) CheckPlan(ctx context.Context, req plancheck.CheckPlanRequest, resp *plancheck.CheckPlanResponse) {
-	rd, err := json.Marshal(req.Plan)
-	if err != nil {
-		tflog.Debug(ctx, fmt.Sprintf("error marshaling machine-readable plan output: %s", err))
-	}
-	tflog.Info(ctx, fmt.Sprintf("req.Plan - %s\n", string(rd)))
-}
-
-func DebugPlan() plancheck.PlanCheck {
-	return debugPlan{}
 }
